more aggressive termination of light propagation thread, rename a doc folder, CLSupportingLevel to detect which levels should and should not support colored lighting
