+++ docs/shader/autopatcher.md
