+++ src/main/java/me/erykczy/colorfullighting/api/CLSupportingLevel.java
@@ -1,0 +1,4 @@
+package me.erykczy.colorfullighting.api;
+
+public interface CLSupportingLevel {
+}
