+++ src/main/java/me/erykczy/colorfullighting/mixin/ClientLevelMixin.java
@@ -1,5 +1,6 @@
 package me.erykczy.colorfullighting.mixin;
 
+import me.erykczy.colorfullighting.api.CLSupportingLevel;
 import me.erykczy.colorfullighting.common.accessors.mixin.ClientLevelAccessor;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.LevelRenderer;
@@ -8,7 +9,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 
 @Mixin(ClientLevel.class)
-public class ClientLevelMixin implements ClientLevelAccessor {
+public class ClientLevelMixin implements ClientLevelAccessor, CLSupportingLevel {
 	@Shadow
 	@Final
 	private LevelRenderer levelRenderer;
